Generalize Config -> Preferences + storage engine.
Generalize some things so that we will be able to run on more than one platform (e.g. Android).

We need to be able to wrap up platform-specific preference storage (Properties files on the desktop vs. SharedPreferences on Android). Also, there's just no reason for things like storage directories to be hardcoded in two places.