Actually install the BouncyCastle crypto provider.
This is a bit of a no-brainer, but the current approach just makes me look like I have no brain. However, since we are going to need some frightfully clever dependency injection stuff when we start up anyway, I'm ok with kicking the elegance can down the road a little ways.