Remove KernelInterface mock.
We can't mock the KernelInterface in the bootstrap integration test because:
1. KernelInterface must be loaded by the FootlightsClassLoader in order to be considered the same type, and
2. Mockito can only mock classes that are visible to it at compile time.

So, it looks like we'll just have to live with an integration test which loads and initializes the Plugin, as well as ensuring that it does something useful, but which doesn't make any kernel calls. That's fine by me: this isn't the most appropriate place to do such unit tests.