Proxy DOM event handlers properly.
It's unsafe to let sandboxed code install event handlers directly: when element.onclick() runs, 'this' is a reference to the naked DOM node. Instead, we need to proxy event handler execution so that 'this' can be a reference to the proxy.

We do this by storing code in the proxy, named by itself. Thus, proxy[<code>] is a compiled version of <code>, and if we run proxy[<code>](), 'this' is defined to what we want.

Finally, we construct a custom environment object out of a sandbox context's global variables, making all global variables (including the special variable 'context') available to proxied event handlers.

This commit fixes Issue #2.