Add the TicTacToe plugin.
This is a demo plugin that will do meaningfully interesting things, such as draw interactive graphics, display proxied user information and talk to the Internet so that players can compete online. All of this should happen without revealing user data to the plugin.

Oh, and the plugin should actually play tic-tac-toe correctly, which it doesn't yet.