Let an EncryptedBlock know about its plaintext.
We need a canonical representation somewhere (does A correspond to B or vice versa?), so here it is: an EncryptedBlock object (which already has a reference to a Link that can decrypt it) should know about its own plaintext Block. Settled.

Now, when we build a local representation of a remote block, which might contain very special "random" padding, we treat the ciphertext as the canonical object, and the plaintext as a mere derivation of it.