Supply Preferences to plugins on initialization.
This is the beginning of persistent storage for plugins (Issue #3): plugins will get a set of key-value pairs on init, some of which could be filenames that they can now ask to kernel to open on their behalf.

There's nothing yet about how plugins will store these persistent preferences, but that's an exercise for later.