Let a link have a Fingerprint, rather than a URI.
Links to content should neither know nor care *where* that content is stored; they should only care about that content's name, which must be a function of the content itself.

This simplifies the implementation a little bit, makes it easy/possible for a Link to store its (Fingerprint,SecretKey) pair in the Keychain and---most importantly---lets us stop giving out the key bytes to anyone who asks.

Finally, use PowerMock to mock out static Fingerprint methods and better isolate the code that we're actually testing.