Remove intermediate Flusher Builder.
Instead of building an intermediate Builder, which accepts optional parameters such as a timeout, provide an explicit wait function which waits however long is correct for the kind of thing that we're flushing.

For the two kinds of Flusher that we currently deal with (Flushable and HasBytes=>File), base this wait function on monitor notifications.