Fix File.getInputStream().
Previously, we didn't correctly handle clients asking for smaller chunks of data than remained in available buffers. As it turns out, object deserialization likes to pull small chunks of header data out first, but at any rate, we really ought to be handling this case.

Combined with a bunch of previous work, this finally closes Issue #3.