A Scala decomposition of the Footlights Core.
The Scala version of the kernel is much more flexible, allowing me to "instantiate a kernel with a filesystem and plugin manager, which uses Swing-based powerboxes and whose operations are invoked with JVM privilege".

An alternative implementation shakes out a couple of bugs, such as the fact that we're relying on Java object serialization for storing key-value mappings (!).

Hopefully this will be much more testable that the existing stuff, too, but this is just a first effort.