Provide JVM-side placeholder infrastructure.
Provide all of the Ajax-ey code required to populate placeholders. Note that this doesn't populate them with *correct* data (it's currently a string which is literally like "(function(){placeholder value for \'user.name\'})();"), but we're one bridge further than we were.

Since we're letting plugin JavaScript do Ajax with custom callbacks, we seriously need to re-evaluate the whole "an AjaxHandler must return JavaScript" thing.