Let the root context play by sandboxed rules.
The root context is still a privileged sandbox (we pass in a reference to the 'window' object!), but there's no reason to call document.method() when a proxied method will do just as well.

Now we have less duplication of setting up sandboxes, and we use the same API that we expect plugin authors to use.