Add a static name resolver.
Static names are long-lived URLs which contain mutable references to immutable content. For instance, I could use the URL "http://footlights.me/~jon/work" to point to a particular file in the CAS which represents the current root of my work identity.

The content of such URLs is a JSON object with, at minimum, a "fingerprint" member.

One neat property of using URLs as names is that you don't actually have to resolve it yourself: you could rely on a proxy service telling you that "this URL resolves to that CAS entry (or at least it did 90 minutes ago)".