Move Preferences to Scala, use nice features.
In particular, return None[T] rather than throwing a NoSuchElementException whenever the Preferences don't contain such an element. This should be more efficient, and it definitely produces cleaner, easier-to-read code.

All consumers of Preferences now need an additional get() call, which will blow up in the usual way (NoSuchElementException) if it fails. So, while the API is changing, it's not really changing. :)

Finally, along the way, there were one or two refactorings that really, really needed to happen with the switch to Scala.