Implement primitive placeholder support.
We don't actually have any placeholders to display yet, but this code would display 'em if we did.

Issue #4 isn't quite dead yet: we still need a way to map names other than "user" to meaningful things (e.g. the Tic-Tac-Toe app might map "opponent" to "profile data of the other player"), but this is a start.