Set up the async channel again when firing events.
This restricts the kind of events that we can fire to just JavaScript. That's ok though, at least for now, since I can't think of any other events that we might like to fire. We might do more interesting things in the future (e.g. "if you ever receive unsolicited JSON, you know what to do with it"), but I'm not sure that would even be a good idea.