Missed this somewhere...
We only need to format three fields in this String, so I'm not sure when we got out of sync, other than I know it was out of sync in the last pushed-to-Github version. I'm kind of surprised that Scala didn't flag this up!

Thus, we no longer need to know the class name of the app (we only care about full names) and we no longer need to import JavaScript.sanitizeText (since we only call it once).