Concentrate usage of sudo().
ClassLoader code and the like should only be called from already-privileged code like the Bootloader and the Kernel; we should not need to do privilege elevation ourselves.

For the moment, we do: scala libraries, unlike core Java libraries, break the chain of privilege in a stack trace, so we need to sudo() inside of maps, etc. Hopefully we will be able to find a remedy for this in the near future, however.