Make application loading more explicit.
Rather than overloading the class namespace, create FootlightsClassLoader.loadApplication() and make it visible in the Kernel through reflection.

Applications are now incapable of asking the bootloader to load other applications (this would've failed anyway, due to a lack of permissions, but the new way is cleaner and easier to reason about).