Load an application by "main" class.
Instead of specifying the main class of an application, use the manifest-based mechanism to only start executing in the class that the application author intended.

This reduces the burden on app startup (one can say, "start this app" rather than "start class X within this app") and provides consistent execution (all users will start running the app in the same place).

This also means that we must change the way that we name applications inside the Web UI: instead of class name (which we no longer use), we should use whatever name the AppWrapper gives us. Of course, since we don't control the content of that name, we must be careful to URLEncode and URLDecode it so as to not confuse slashes in the name with slashes that separate elements of the WebRequest.