Express dependencies as URIs, not URLs.
As we move towards actually using dependencies, we will start having two kinds: absolute hash values (which are URIs with an unofficial scheme but not URLs, since they give no location information) and indirection links (which are URLs that redirect to hash-value URIs).

To be fully general, change dependency URLs into URIs.