Bring Android POM into the modern world.
With the latest android-maven-plugin, we can run ProGuard fairly painlessly (i.e. with a LOT less configuration than we used to need), which lets us start code that uses Scala.

The Kernel wants to be refactored a little, making it easier to use on both the desktop and Android. In the meantime, just comment the kernel out of the Android app so that the app will run.

The ProGuard configuration is non-standard: we absolutely need ProGuard to "shake the tree" in order to stay under the 64k method limit, but we don't want it obfuscating or optimising away any dependency injection code.

We also update to modern RoboGuice (which requires a new Android Compatibility library), specify our preferred version of Guice, etc.