Move ScalaTest dependency to root POM.
Since the artifactId embeds the Scala ABI number, it will be a pain to change to Scala 2.9.2 if it's scattered all over a ton of POMs. If, on the other hand, we put it in one place, it will be easier.

This does mean adding an unnecessary dependency to some projects that don't use ScalaTest, but it's scoped to the test phase, so it shouldn't cause any compile- or run-time bloat.