Fix up some poor class loading assumptions.
There's no reason why two JAR files can't contain files from the same package (indeed, this is what some Apache stuff we'd like to use does).

Instead of keeping a package->ClasspathLoader mapping in the FootlightsClassLoader and building an iterator over possible ClasspathLoaders every time we load a class (believe it or not, this was originally meant as a performance optimisation), construct all ClasspathLoaders on init.

This means that a ClasspathLoader can't know ahead of time what its "base" package is. So, we need to adjust the "do I consult mummy" logic. Split the "must consult parent" case into "must" (language runtime) and "may" (Footlights core) cases: the second says that findClass() is allowed to ask mummy if we can't find a class ourselves. This also requires making a bit of FootlightsClassLoader visible to ClasspathLoader.