Add Directory support.
A directory is a mapping of names to entries, where an entry may either be a file or another directory.

A *mutable* directory (which includes the Directory interface now exposed through the Footlights API) is a wrapper around an immutable directory which knows how to notify some parent when it changes.