Use Either where appropriate in KernelInterface.
Rather than returning Option (which provides no debug information) and possibly throwing an exception (which provides lots), use the Scala Either class to indicate "usually this, but possibly that Exception".

Co-mingled with this change is a check in Directory.parse() to make sure that we aren't trying to parse an empty list of blocks.