Remove private keys from Keychain.
Keychain has two purposes:
1. to store keys and
2. to interact with Java KeyStores on disk.

We can't deal sensibly with private keys in a JKS: it brings in all kinds of useless overhead around x509 certificates, etc. So, we'll just store public/private keys in the Footlights CAS.

The root of all this trust, when running on a local computer, was always going to be trusted computer + password. Now, we'll just stop storing the user's Ultimate Private Key in a Java KeyStore.