Convert SwingUI main class to Scala.
This will become important in a bit, when I define some new UI responsibilities (e.g. prompting the user to enter data or choose an option, then calling a Scala callback). It's not impossible to do this kind of interop with pure Java, but it'd be a bit ugly and possibly take more time than doing this conversion.