Create the data to be shared with another user.
Put all information which has been shared with another user into a directory, the root of which can be shared.

This sharing of a root directory hasn't actually been done yet.

Also, return the URI of the other user (a fingerprint of their public key). This provides an opaque but stable handle which can be used to reference the other user.