Fix the "blocking flush" problem.
Instead of synchronising access to a mutable journal, synchronise access to a variable which contains an immutable journal.

It is always safe to iterate over a snapshot of an immutable Set, so we don't need to put the slow I/O stuff in the critical section. When we've flushed some number of blocks, atomically remove the flushed blocks from the journal.

This speeds things up a *lot*.