localizeJar returns an Either now, not an Option.
This commit fixes library loading.

I changed localizeJar() awhile ago to return an Either, not an Option, but didn't update the Bootstrapper consumer. Bootstrapper would just print out messages like "I've received something that isn't an Option".

This is the danger of loose coupling, but there's not much that can be done about it: the whole point of the Bootstrapper is to couple loosely so that we don't get circular dependencies in class loading.